err filter realisasi kpi

+++ database/migrations/2025_01_28_152636_create_kpis_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->string('indikator_kinerja');
             $table->integer('id_kategori');
             $table->integer('id_satuan');
-            $table->integer('bobot');
+            $table->double('bobot');
             $table->integer('polaritas');
             $table->timestamps();
         });
